chhore:ensure-i64-used-in-query-params

+++ src/apis/position_api.rs
@@ -36,7 +36,7 @@ pub struct PositionControllerListBySubaccountIdParams {
     /// Direction to paginate through objects
     pub order: Option<String>,
     /// Limit the number of objects to return
-    pub limit: Option<i32>,
+    pub limit: Option<i64>,
     /// Pointer to the current object in pagination dataset
     pub cursor: Option<String>,
     /// Array of product ids to filter for
@@ -46,9 +46,9 @@ pub struct PositionControllerListBySubaccountIdParams {
     /// Order by field
     pub order_by: Option<String>,
     /// Filter by order fills created before timestamp exclusive (ms since Unix epoch)
-    pub created_after: Option<i32>,
+    pub created_after: Option<i64>,
     /// Filter by order fills created before timestamp inclusive (ms since Unix epoch)
-    pub created_before: Option<i32>,
+    pub created_before: Option<i64>,
     /// Side as either BUY (0) or SELL (1)
     pub side: Option<f64>,
     /// Filter by liquidated positions
@@ -63,7 +63,7 @@ pub struct PositionControllerListFillsByPositionIdParams {
     /// Direction to paginate through objects
     pub order: Option<String>,
     /// Limit the number of objects to return
-    pub limit: Option<i32>,
+    pub limit: Option<i64>,
     /// Pointer to the current object in pagination dataset
     pub cursor: Option<String>,
     /// Order by field
@@ -76,7 +76,7 @@ pub struct PositionControllerListLiquidationsBySubaccountIdParams {
     /// Direction to paginate through objects
     pub order: Option<String>,
     /// Limit the number of objects to return
-    pub limit: Option<i32>,
+    pub limit: Option<i64>,
     /// Pointer to the current object in pagination dataset
     pub cursor: Option<String>,
     /// Order by field
